pyproject.toml
[project]
name = "subtranslator-ng-roo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "srt>=3.5.3",
    "werkzeug>=3.0.0",
    "tenacity>=8.2.3",
    "google-genai>=1.0.0",
    "flask>=3.1.0",
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "aiofiles>=24.1.0",
    "pytest>=8.3.5",
]
